+++ testuser.csv
@@ -1,0 +1,1 @@
+"0","A","","0-0-0","T","","","000","000","","0",""
